Finished Level and Goal classes and added tag functionality to Body
The Level class now has most, if not all of the functionality it needs to work within the game, it will create a level with given parameters via the Level::newLevel() function as well as keep everything up to date, further abstracting from the Game class.The Goal class now exists as a Body that is a sensor and is b2Static. I also added in a tag to the body class so that, even though collisions occur and return pointers to Body objects, you can still tailor the response by checking what tag the body has. I have used this on the Ball class to check if the body it collided with was tagged as a Goal, and if so, destroy itself and score a point