Fixed issues, and game is playable
Fixed the big issue of balls sort of floating through platforms, I was rotating the platform in radians when i should have be using degrees. The game is in a sort of playable state with a default level without platform placement, balls fall into goals, if enough balls fall into a goal, you win.